requirements.txt
tensorflow>=2.10.0
numpy
pandas
opencv-python
pyyaml
tqdm
matplotlib
google-generativeai
scikit-learn